requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
WTForms==3.1.2
PyJWT==2.8.0
bcrypt==4.1.2
pillow==10.3.0
opencv-python==4.9.0.80
gunicorn==23.0.0